Add more DAO methods and add junit test to test DAO saveBooking methods.

+++ adapters/custody-analytics-staging-adapter/src/main/java/org/ojbc/adapters/analyticsstaging/custody/dao/model/Booking.java
@@ -41,7 +41,7 @@ public class Booking implements Serializable{
 	private Integer bedTypeId; 
 	private BigDecimal arrestLocationLatitude; 
 	private BigDecimal arrestLocationLongitude; 
-	private Long bookingSubjectId; 
+	private Integer bookingSubjectId; 
 	
 	public Integer getBookingId() {
 		return bookingId;
@@ -151,11 +151,11 @@ public class Booking implements Serializable{
 		this.arrestLocationLongitude = arrestLocationLongitude;
 	}
 
-	public Long getBookingSubjectId() {
+	public Integer getBookingSubjectId() {
 		return bookingSubjectId;
 	}
 
-	public void setBookingSubjectId(Long bookingSubjectId) {
+	public void setBookingSubjectId(Integer bookingSubjectId) {
 		this.bookingSubjectId = bookingSubjectId;
 	}
